requirements.txt
Flask
pyyaml
attrdict
pymongo
